refactor to use better structure for tests
