+++ tests/Unit/Countries/US/NorthCarolina/NorthCarolinaUnemployment/V20180101/NorthCarolinaUnemploymentTest.php
@@ -1,17 +1,18 @@
 <?php
 
-namespace Appleton\Taxes\Countries\US\NorthCarolina\NorthCarolinaUnemployment;
+namespace Appleton\Taxes\Countries\US\NorthCarolina\NorthCarolinaUnemployment\V20180101;
 
 use Carbon\Carbon;
+use Appleton\Taxes\Countries\US\NorthCarolina\NorthCarolinaUnemployment\NorthCarolinaUnemployment as ParentNorthCarolinaUnemployment;
 
-class NorthCarolinaUnemploymentV20190101Test extends \TestCase
+class NorthCarolinaUnemploymentTest extends \TestCase
 {
     public function setUp()
     {
         parent::setUp();
 
         Carbon::setTestNow(
-            Carbon::parse('January 1, 2019 8am', 'America/Chicago')->setTimezone('UTC')
+            Carbon::parse('January 1, 2018 8am', 'America/Chicago')->setTimezone('UTC')
         );
     }
 
@@ -24,7 +25,7 @@ class NorthCarolinaUnemploymentV20190101Test extends \TestCase
             $taxes->setEarnings(2300);
         });
 
-        $this->assertSame(23.00, $results->getTax(NorthCarolinaUnemployment::class));
+        $this->assertSame(23.00, $results->getTax(ParentNorthCarolinaUnemployment::class));
     }
 
     public function testNorthCarolinaUnemploymentWithTaxRate()
@@ -38,7 +39,7 @@ class NorthCarolinaUnemploymentV20190101Test extends \TestCase
             $taxes->setEarnings(2300);
         });
 
-        $this->assertSame(55.20, $results->getTax(NorthCarolinaUnemployment::class));
+        $this->assertSame(55.20, $results->getTax(ParentNorthCarolinaUnemployment::class));
     }
 
     public function testNorthCarolinaUnemploymentMetWageBase()
@@ -48,40 +49,40 @@ class NorthCarolinaUnemploymentV20190101Test extends \TestCase
             $taxes->setWorkLocation($this->getLocation('us.north_carolina'));
             $taxes->setUser($this->user);
             $taxes->setEarnings(100);
-            $taxes->setYtdEarnings(24200);
+            $taxes->setYtdEarnings(23300);
         });
 
-        $this->assertSame(1.00, $results->getTax(NorthCarolinaUnemployment::class));
+        $this->assertSame(1.00, $results->getTax(ParentNorthCarolinaUnemployment::class));
 
         $results = $this->taxes->calculate(function ($taxes) {
             $taxes->setHomeLocation($this->getLocation('us.north_carolina'));
             $taxes->setWorkLocation($this->getLocation('us.north_carolina'));
             $taxes->setUser($this->user);
             $taxes->setEarnings(100);
-            $taxes->setYtdEarnings(24250);
+            $taxes->setYtdEarnings(23450);
         });
 
-        $this->assertSame(0.50, $results->getTax(NorthCarolinaUnemployment::class));
+        $this->assertSame(0.50, $results->getTax(ParentNorthCarolinaUnemployment::class));
 
         $results = $this->taxes->calculate(function ($taxes) {
             $taxes->setHomeLocation($this->getLocation('us.north_carolina'));
             $taxes->setWorkLocation($this->getLocation('us.north_carolina'));
             $taxes->setUser($this->user);
             $taxes->setEarnings(100);
-            $taxes->setYtdEarnings(24300);
+            $taxes->setYtdEarnings(23500);
         });
 
-        $this->assertSame(0.0, $results->getTax(NorthCarolinaUnemployment::class));
+        $this->assertSame(0.0, $results->getTax(ParentNorthCarolinaUnemployment::class));
 
         $results = $this->taxes->calculate(function ($taxes) {
             $taxes->setHomeLocation($this->getLocation('us.north_carolina'));
             $taxes->setWorkLocation($this->getLocation('us.north_carolina'));
             $taxes->setUser($this->user);
             $taxes->setEarnings(100);
-            $taxes->setYtdEarnings(24350);
+            $taxes->setYtdEarnings(23550);
         });
 
-        $this->assertSame(0.0, $results->getTax(NorthCarolinaUnemployment::class));
+        $this->assertSame(0.0, $results->getTax(ParentNorthCarolinaUnemployment::class));
     }
 
     public function testCaseStudy1()
@@ -95,7 +96,7 @@ class NorthCarolinaUnemploymentV20190101Test extends \TestCase
             $taxes->setEarnings(66.68);
         });
 
-        $this->assertSame(1.27, $results->getTax(NorthCarolinaUnemployment::class));
+        $this->assertSame(1.27, $results->getTax(ParentNorthCarolinaUnemployment::class));
     }
 
     public function testWorkOutOfState()
@@ -110,6 +111,6 @@ class NorthCarolinaUnemploymentV20190101Test extends \TestCase
         });
 
         $this->assertNull($results->getTax(NorthCarolinaIncome::class));
-        $this->assertSame(1.27, $results->getTax(NorthCarolinaUnemployment::class));
+        $this->assertSame(1.27, $results->getTax(ParentNorthCarolinaUnemployment::class));
     }
 }
